% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{check_currency_symbol}
\alias{check_currency_symbol}
\title{Checks whether currency pair is supported.}
\usage{
check_currency_symbol(pair)
}
\arguments{
\item{pair}{}
}
\value{

}
\description{
Checks whether currency pair is supported.
}
\examples{
\dontrun{
check_currency_symbol("GBP/USD")
check_currency_symbol("USD/GBP")
}
}
