% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{check_token}
\alias{check_token}
\title{Check the polygon token environment variable.}
\usage{
check_token()
}
\description{
Check the polygon token environment variable.
}
\details{
Check that the POLYGON_TOKEN environment variable has been set.
You can obtain an API key from the [Polygon.io Website](https://polygon.io/).
}
