% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{clean_response}
\alias{clean_response}
\title{create_friendly_names}
\usage{
clean_response(response)
}
\arguments{
\item{response}{A response object.}
}
\value{
A tibble.
}
\description{
create_friendly_names
}
\keyword{internal}
