% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_friendly_names}
\alias{create_friendly_names}
\title{create_friendly_names}
\usage{
create_friendly_names(data)
}
\arguments{
\item{data}{A tibble.}
}
\value{
A tibble.
}
\description{
create_friendly_names
}
\keyword{internal}
