% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_aggregates}
\alias{get_aggregates}
\title{get_aggregates}
\usage{
get_aggregates(token, ticker, multiplier, timespan, from, to)
}
\arguments{
\item{token}{A character string of an appropriate Ticker symbol of the request.}

\item{ticker}{An integer for the size of the timespan multiplier.}

\item{multiplier}{A character string  of the size of the time window.
Options include: 'minute', 'hour', 'day', 'week', 'month', 'quarter', 'year'.}

\item{to}{}
}
\value{
A dataframe.
}
\description{
Get aggregates for a date range, in custom time window sizes.
}
\examples{
\dontrun{

libraray(polygon)

get_aggregates(
token = "YOUR_POLYGON_TOKEN",
ticker = "AAPL",
multiplier = 1,
timespan = "day",
from = "2019-01-01",
to = "2019-02-01"
)
}

}
