% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_exchanges}
\alias{get_exchanges}
\title{get_exchanges}
\usage{
get_exchanges(token)
}
\arguments{
\item{token}{(string) A valid token for polygonio.}
}
\value{
A tibble.
}
\description{
List of stock exchanges which are supported by Polygon.io
}
\examples{
\dontrun{
library(polygon)
get_exchanges(token)
}
}
