% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypto.R
\name{get_exchanges_cypto}
\alias{get_exchanges_cypto}
\title{get_exchanges_cypto}
\usage{
get_exchanges_cypto(token)
}
\arguments{
\item{token}{(string) A valid token for polygonio.}
}
\value{
A tibble.
}
\description{
Get list of crypto currency exchanges which are supported by Polygon.io
}
\examples{
\dontrun{
library(polygon)
get_exchanges_cypto(token)
}
}
