% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_gainers_and_loser}
\alias{get_gainers_and_loser}
\title{get_gainers_and_loser}
\usage{
get_gainers_and_loser(token, direction = c("gainers", "losers"))
}
\arguments{
\item{token}{(string) A valid token for polygonio.}

\item{direction}{(string) Direction we want. Options include: "gainers" and
"losers".}
}
\value{
A tibble.
}
\description{
Get the current snapshot of the top 20 gainers or losers
of the day.
}
\examples{
\dontrun{
library(polygon)
token = "YOUR_POLYGON_TOKEN",
get_gainers_and_loser(token, "gainers")
}
}
