% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_grouped_daily_bars}
\alias{get_grouped_daily_bars}
\title{get_grouped_daily_bars}
\usage{
get_grouped_daily_bars(token, locale = c("G", "US", "GB", "CA", "NL",
  "GR", "SP", "DE", "BE", "DK", "FI", "IE", "PT", "IN", "MX", "FR", "CN",
  "CH", "SE"), market = c("STOCKS", "CRYPTO", "BONDS", "MF", "MMF",
  "INDICES", "FX"), date)
}
\arguments{
\item{token}{(string) A valid token for polygonio.}

\item{locale}{(string) Locale of the aggregates. Run `get_locales()` to get
a list of all possible locales.}

\item{market}{(integer) Market of the aggregates. Run `get_locales()` to get
a list of all possible locales.
options include: "STOCKS", "CRYPTO", "BONDS", "MF", "MMF", "INDICES", "FX".}

\item{date}{(string) From date ('YYYY-MM-DD').}
}
\value{
A tibble of financial data.
}
\description{
Get the daily OHLC for entire markets..
}
\examples{
\dontrun{
library(polygon)
get_aggregates(
token = "YOUR_POLYGON_TOKEN",
ticker = "AAPL",
multiplier = 1,
timespan = "day",
from = "2019-01-01",
to = "2019-02-01"
)
}
}
