% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_historic_quotes}
\alias{get_historic_quotes}
\title{get_historic_quotes}
\usage{
get_historic_quotes(token, ticker, date)
}
\arguments{
\item{token}{(string) A valid token for polygonio.}

\item{ticker}{(string) An appropriate Ticker}

\item{date}{(string) Date of the historic ticks to retrieve ('YYYY-MM-DD').}
}
\value{
A tibble of financial data.
}
\description{
Get historic NBBO quotes for a ticker.
}
\examples{
\dontrun{
library(polygon)

get_historic_quotes(
token = "YOUR_POLYGON_TOKEN",
ticker = "AAPL",
date = "2019-01-01"
)
}
}
