% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_locales}
\alias{get_locales}
\title{get_locales}
\usage{
get_locales()
}
\value{
A tibble.
}
\description{
Get the list of currently supported locales
}
\examples{
\dontrun{
library(polygon)
get_locales()
}
}
