% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_news}
\alias{get_news}
\title{Get news from polygon API}
\usage{
get_news(ticker, publish_date = Sys.Date() - 1, limit = 100)
}
\arguments{
\item{ticker}{(string) Stock ticker symbol of interest.}

\item{publish_date}{(Date) Publish date. Return articles publish at or
after this date.}

\item{limit}{(Integer) The maximum number of results returned.}
}
\description{
Get the most recent news articles relating to a
stock ticker symbol, including a summary of the article
and a link to the original source.
}
\examples{
\dontrun{
get_news('AAPL')
}
}
