% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_previous_close}
\alias{get_previous_close}
\title{get_previous_close}
\usage{
get_previous_close(ticker)
}
\arguments{
\item{ticker}{A character string of an appropriate Ticker.}
}
\value{
A tibble of financial data.
}
\description{
Get the previous day close for the specified ticker
}
\examples{
\dontrun{
library(polygon)

get_previous_close(
ticker = "AAPL"
)
}
}
