% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypto.R
\name{get_snapshot_all_tickers_cypto}
\alias{get_snapshot_all_tickers_cypto}
\title{get_snapshot_all_tickers_cypto}
\usage{
get_snapshot_all_tickers_cypto(token)
}
\arguments{
\item{token}{A valid token for polygonio (character string).}

\item{ticker}{A character string of an appropriate Ticker.}
}
\value{
A tibble of financial data.
}
\description{
Snapshot allows you to see all Crypto tickers' current
minute aggregate, daily aggregate and last trade.
As well as previous days aggregate and calculated change for today.
}
\examples{
\dontrun{
library(polygon)

get_snapshot_all_tickers_cypto(
token = "YOUR_POLYGON_TOKEN",
ticker = "X:BTCUSD"
)
}
}
