% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_snapshot_all_tickers_stocks}
\alias{get_snapshot_all_tickers_stocks}
\title{get_snapshot_all_tickers_stocks}
\usage{
get_snapshot_all_tickers_stocks(token, ticker)
}
\arguments{
\item{token}{A valid token for polygonio (character string).}

\item{ticker}{A character string of an appropriate Ticker.}
}
\value{
A tibble of financial data.
}
\description{
Snapshot allows you to see all Stock tickers' current
minute aggregate, daily aggregate and last trade.
As well as previous days aggregate and calculated change for today.
Warning, the response size can be relatively large.
}
\examples{
\dontrun{
library(polygon)

get_snapshot_all_tickers_stocks(
token = "YOUR_POLYGON_TOKEN",
ticker = "AAPL"
)
}
}
