% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_forex}
\alias{historic_forex}
\title{Retrieve historic data for a currency pair}
\usage{
historic_forex(pair, date, limit = 100)
}
\arguments{
\item{pair}{Currency pair}

\item{date}{Date}

\item{limit}{Number of records to retrieve}
}
\value{
A data frame
}
\description{
Retrieve historic data for a currency pair
}
\examples{
\dontrun{
historic_forex("AUD/USD", "2018-02-02", limit = 20)
}
}
