% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last-quote.R
\name{last_quote_forex}
\alias{last_quote_forex}
\title{Retrieve last quote for a currency pair}
\usage{
last_quote_forex(pair)
}
\arguments{
\item{pair}{Currency pair}
}
\value{
A data frame
}
\description{
Retrieve last quote for a currency pair
}
\examples{
\dontrun{
last_quote_forex("AUD/USD")
}
}
