% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon-package.R
\docType{package}
\name{polygon-package}
\alias{polygon-package}
\alias{_PACKAGE}
\title{polygon: R Websocket & RESTful API client for Polygon}
\description{
Provides an R WebSocket and RESTful API client for Polygon.
    Polygon is a platform for real-time and historical financial market data
    including stocks APIs, forex, crypto, and more: <https://polygon.io/>.
}
\author{
\strong{Maintainer}: Ernest Omane-Kodie \email{eokodie@gmail.com}

}
\keyword{internal}
