% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{polygon_auth}
\alias{polygon_auth}
\title{Polygon Authentication}
\usage{
polygon_auth()
}
\description{
`polygon_auth()` looks for the environment variable `POLYGON_TOKEN`.
We recommend that you set this variable in `.Renviron`.
}
