% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{posix_to_timestamp}
\alias{posix_to_timestamp}
\title{Converts POSIXct to numeric timestamp}
\usage{
posix_to_timestamp(posix)
}
\arguments{
\item{posix}{POSIXct}
}
\value{

}
\description{
Converts POSIXct to numeric timestamp
}
