% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate.R
\name{set_api_key}
\alias{set_api_key}
\title{Title}
\usage{
set_api_key(api_key)
}
\arguments{
\item{api_key}{API key}
}
\value{
None
}
\description{
Title
}
\examples{
set_api_key("KkbzopFGTg3PMv_CNqaMRUfdhiSJ7rzdi9DELl")
}
