% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp}
\alias{timestamp}
\title{Converts numeric timestamp into POSIXct}
\usage{
timestamp(msec)
}
\arguments{
\item{msec}{milliseconds since the epoch}
}
\value{
POSIXct
}
\description{
Converts numeric timestamp into POSIXct
}
\examples{
timestamp(1517529600225)
}
