% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timestamp.R
\name{timestamp_to_posix}
\alias{timestamp_to_posix}
\title{Converts numeric timestamp into POSIXct}
\usage{
timestamp_to_posix(msec)
}
\arguments{
\item{msec}{milliseconds since the epoch}
}
\value{
POSIXct
}
\description{
Converts numeric timestamp into POSIXct
}
\examples{
timestamp_to_posix(1517529605225)
}
