% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\name{WebSocket}
\alias{WebSocket}
\title{A WebSocket template}
\value{
A WebSocket template.
}
\description{
This class manages WebSocket connections to polygon.io clusters.
}
\section{Public Methods}{

\describe{
  \item{\code{close}}{close WebSocket connection.}
  \item{\code{connect_to_cluster}}{connect to a polygon WebSocket cluster.}
  \item{\code{authenticate}}{Authenticate connection.}
  \item{\code{subscribe}}{Subscribe to stream.}
  \item{\code{unsubscribe}}{Unsubscribe from a stream.}
}
}

\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{WebSocket$new()}}
\item \href{#method-close}{\code{WebSocket$close()}}
\item \href{#method-connect_to_cluster}{\code{WebSocket$connect_to_cluster()}}
\item \href{#method-authenticate}{\code{WebSocket$authenticate()}}
\item \href{#method-subscribe}{\code{WebSocket$subscribe()}}
\item \href{#method-unsubscribe}{\code{WebSocket$unsubscribe()}}
\item \href{#method-finalize}{\code{WebSocket$finalize()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$new(cluster, token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-close"></a>}}
\if{latex}{\out{\hypertarget{method-close}{}}}
\subsection{Method \code{close()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-connect_to_cluster"></a>}}
\if{latex}{\out{\hypertarget{method-connect_to_cluster}{}}}
\subsection{Method \code{connect_to_cluster()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$connect_to_cluster(cluster)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-authenticate"></a>}}
\if{latex}{\out{\hypertarget{method-authenticate}{}}}
\subsection{Method \code{authenticate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$authenticate(token)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subscribe"></a>}}
\if{latex}{\out{\hypertarget{method-subscribe}{}}}
\subsection{Method \code{subscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$subscribe(tickers)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unsubscribe"></a>}}
\if{latex}{\out{\hypertarget{method-unsubscribe}{}}}
\subsection{Method \code{unsubscribe()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$unsubscribe(tickers)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WebSocket$finalize()}\if{html}{\out{</div>}}
}

}
}
