% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_http_status}
\alias{check_http_status}
\title{check_http_status}
\usage{
check_http_status(response)
}
\arguments{
\item{data}{A request object.}
}
\description{
Check http response messages.
}
\keyword{internal}
