% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_aggregates}
\alias{get_aggregates}
\title{get_aggregates}
\usage{
get_aggregates(
  ticker,
  multiplier,
  timespan = c("minute", "hour", "day", "week", "month", "quarter", "year"),
  from,
  to
)
}
\arguments{
\item{ticker}{(string) Ticker symbol of the request.
Some tickers require a prefix, see examples below:
Stocks: 'AAPL'
Currencies: 'C:EURUSD'
Crypto: 'X:BTCUSD'}

\item{multiplier}{(integer) Size of the timespan multiplier.}

\item{timespan}{(string) Size of the time window.
Options include: 'minute', 'hour', 'day', 'week', 'month',
'quarter', 'year'.}

\item{from}{(Date) From date ('YYYY-MM-DD').}

\item{to}{(Date) To date ('YYYY-MM-DD').}
}
\value{
A tibble of financial data.
}
\description{
Get OHLC aggregates for a date range, in custom
time window sizes for a ticker.
}
\examples{
\dontrun{
library(polygon)
get_aggregates(
ticker = "AAPL",
multiplier = 1,
timespan = "day",
from = "2019-01-01",
to = "2019-02-01"
)
}
}
