% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_exchanges}
\alias{get_exchanges}
\title{get_exchanges}
\usage{
get_exchanges(token)
}
\arguments{
\item{token}{(string) A valid token for polygonio.}
}
\value{
A tibble.
}
\description{
Get list of stock exchanges which are supported by Polygon.io
}
\examples{
\dontrun{
library(polygon)
token = "YOUR_POLYGON_TOKEN",
get_exchanges(token)
}
}
