% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_grouped_daily_bars}
\alias{get_grouped_daily_bars}
\title{get_grouped_daily_bars}
\usage{
get_grouped_daily_bars(date, unadjusted = FALSE)
}
\arguments{
\item{date}{(Date) From date ('YYYY-MM-DD').}

\item{unadjusted}{(Logical) Whether or not the results are adjusted
for splits. By default, results are adjusted. Set this to true to
get results that are NOT adjusted for splits.}
}
\value{
A tibble of financial data.
}
\description{
Get the daily OHLC for entire markets..
}
\examples{
\dontrun{
library(polygon)
get_aggregates(
date = Sys.Date() - 1
)
}
}
