% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cypto.R
\name{get_open_close_crypto}
\alias{get_open_close_crypto}
\title{Daily Open/Close for Crypto}
\usage{
get_open_close_crypto(from, to, date)
}
\arguments{
\item{from}{(Date) From Symbol of the pair.}

\item{to}{(Date) To Symbol of the pair.}

\item{date}{(Date) Date of the requested open/close.}
}
\value{
A tibble of financial data.
}
\description{
Get the open, close prices of a symbol on a certain day.
}
\examples{
\dontrun{
library(polygon)
get_open_close_crypto(
from = "BTC",
to = "USD",
date = "2019-01-01"
)
}
}
