% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocks.R
\name{get_snapshot_all_tickers_stocks}
\alias{get_snapshot_all_tickers_stocks}
\title{get_snapshot_all_tickers_stocks}
\usage{
get_snapshot_all_tickers_stocks()
}
\value{
A tibble of financial data.
}
\description{
Snapshot allows you to see all Stock tickers' current
minute aggregate, daily aggregate and last trade.
As well as previous days aggregate and calculated change for today.
}
\examples{
\dontrun{
library(polygon)

get_snapshot_all_tickers_stocks()
}
}
