% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historic.R
\name{historic_forex}
\alias{historic_forex}
\title{Title}
\usage{
historic_forex(pair, date, limit = 100)
}
\arguments{
\item{limit}{}
}
\description{
Title
}
\examples{
\dontrun{
historic_forex("AUD/USD", "2018-02-02", limit = 20)
}
}
