% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/websocket.R
\docType{data}
\name{WebSocket}
\alias{WebSocket}
\title{A WebSocket template}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
WebSocket
}
\value{
A WebSocket template.
}
\description{
This class manages WebSocket connections to polygon.io clusters.
}
\section{Public Methods}{

\describe{
  \item{\code{close}}{close WebSocket connection.}
  \item{\code{connect_to_cluster}}{connect to a polygon WebSocket cluster.}
  \item{\code{authenticate}}{Authenticate connection.}
  \item{\code{subscribe}}{Subscribe to stream.}
  \item{\code{unsubscribe}}{Unsubscribe from a stream.}
}
}

\keyword{datasets}
